+++ src/TreeSearch-init.c
@@ -6,9 +6,6 @@
 #include <stdlib.h> /* for NULL */
 #include <R_ext/Rdynload.h>
 
-#include "mpl.h"
-#include "RMorphyUtils.h"
-#include "RMorphy.h"
 #include "build_postorder.h"
 
 extern SEXP _TreeSearch_nni(SEXP, SEXP, SEXP);
@@ -18,10 +15,6 @@ extern SEXP _TreeSearch_tbr(SEXP, SEXP);
 // extern SEXP _TreeSearch_tbr_moves(SEXP);
 extern SEXP _TreeSearch_all_spr(SEXP, SEXP);
 extern SEXP _TreeSearch_all_tbr(SEXP, SEXP);
-extern SEXP _TreeSearch_preorder_morphy(SEXP, SEXP);
-extern SEXP _TreeSearch_preorder_morphy_by_char(SEXP, SEXP);
-extern SEXP _TreeSearch_morphy_iw(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-extern SEXP _TreeSearch_morphy_profile(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 
 extern SEXP _TreeSearch_expected_mi(SEXP, SEXP);
 extern SEXP _TreeSearch_mi_key(SEXP, SEXP);
@@ -70,28 +63,6 @@ extern SEXP _TreeSearch_ts_collapsed_flags_debug(SEXP, SEXP, SEXP, SEXP, SEXP, S
 extern SEXP _TreeSearch_ts_collapse_pool(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 
 static const R_CallMethodDef callMethods[] = {
-  {"_R_wrap_mpl_new_Morphy",        (DL_FUNC) &_R_wrap_mpl_new_Morphy, 0},
-  {"_R_wrap_mpl_delete_Morphy",     (DL_FUNC) &_R_wrap_mpl_delete_Morphy, 1},
-  {"_R_wrap_mpl_init_Morphy",       (DL_FUNC) &_R_wrap_mpl_init_Morphy, 3},
-  {"_R_wrap_mpl_get_numtaxa",       (DL_FUNC) &_R_wrap_mpl_get_numtaxa, 1},
-  {"_R_wrap_mpl_get_num_charac",    (DL_FUNC) &_R_wrap_mpl_get_num_charac, 1},
-  {"_R_wrap_mpl_attach_symbols",    (DL_FUNC) &_R_wrap_mpl_attach_symbols, 2},
-  {"_R_wrap_mpl_get_symbols",       (DL_FUNC) &_R_wrap_mpl_get_symbols, 1},
-  {"_R_wrap_mpl_attach_rawdata",    (DL_FUNC) &_R_wrap_mpl_attach_rawdata, 2},
-  {"_R_wrap_mpl_set_parsim_t",      (DL_FUNC) &_R_wrap_mpl_set_parsim_t, 3},
-  {"_R_wrap_mpl_get_gaphandl",      (DL_FUNC) &_R_wrap_mpl_get_gaphandl, 1},
-  {"_R_wrap_mpl_set_gaphandl",      (DL_FUNC) &_R_wrap_mpl_set_gaphandl, 2},
-  {"_R_wrap_mpl_set_num_internal_nodes", (DL_FUNC) &_R_wrap_mpl_set_num_internal_nodes, 2},
-  {"_R_wrap_mpl_get_num_internal_nodes", (DL_FUNC) &_R_wrap_mpl_get_num_internal_nodes, 1},
-  {"_R_wrap_mpl_apply_tipdata",     (DL_FUNC) &_R_wrap_mpl_apply_tipdata, 1},
-  {"_R_wrap_mpl_set_charac_weight", (DL_FUNC) &_R_wrap_mpl_set_charac_weight, 3},
-  {"_R_wrap_mpl_get_charac_weight", (DL_FUNC) &_R_wrap_mpl_get_charac_weight, 2},
-  {"_R_wrap_mpl_first_down_recon",  (DL_FUNC) &_R_wrap_mpl_first_down_recon, 4},
-  {"_R_wrap_mpl_first_up_recon",    (DL_FUNC) &_R_wrap_mpl_first_up_recon, 5},
-  {"_R_wrap_mpl_second_down_recon", (DL_FUNC) &_R_wrap_mpl_second_down_recon, 4},
-  {"_R_wrap_mpl_second_up_recon",   (DL_FUNC) &_R_wrap_mpl_second_up_recon, 5},
-  {"_R_wrap_mpl_update_tip",        (DL_FUNC) &_R_wrap_mpl_update_tip, 3},
-  {"_R_wrap_mpl_update_lower_root", (DL_FUNC) &_R_wrap_mpl_update_lower_root, 3},
   {"_TreeSearch_nni",               (DL_FUNC) &_TreeSearch_nni, 3},
   {"_TreeSearch_spr",               (DL_FUNC) &_TreeSearch_spr, 2},
   {"_TreeSearch_all_spr",           (DL_FUNC) &_TreeSearch_all_spr, 2},
@@ -99,11 +70,6 @@ static const R_CallMethodDef callMethods[] = {
   {"_TreeSearch_tbr",               (DL_FUNC) &_TreeSearch_tbr, 2},
   {"_TreeSearch_all_tbr",           (DL_FUNC) &_TreeSearch_all_tbr, 2},
 //  {"_TreeSearch_tbr_moves",         (DL_FUNC) &_TreeSearch_tbr_moves, 1},
-  {"_TreeSearch_preorder_morphy",   (DL_FUNC) &_TreeSearch_preorder_morphy, 2},
-  {"_TreeSearch_preorder_morphy_by_char",   (DL_FUNC) &_TreeSearch_preorder_morphy_by_char, 2},
-  
-  {"_TreeSearch_morphy_iw",         (DL_FUNC) &_TreeSearch_morphy_iw, 7},
-  {"_TreeSearch_morphy_profile",    (DL_FUNC) &_TreeSearch_morphy_profile, 6},
   {"_TreeSearch_expected_mi",       (DL_FUNC) &_TreeSearch_expected_mi, 2},
   {"_TreeSearch_mi_key",            (DL_FUNC) &_TreeSearch_mi_key, 2},
 
@@ -143,9 +109,7 @@ static const R_CallMethodDef callMethods[] = {
   /* ts_stochastic_tbr (9) and ts_parallel_temper (10) removed */
 
   {"_TreeSearch_mc_fitch_scores",    (DL_FUNC) &_TreeSearch_mc_fitch_scores, 2},
-  {"MORPHYLENGTH",                  (DL_FUNC) &MORPHYLENGTH, 4},
   {"RANDOM_TREE",                   (DL_FUNC) &RANDOM_TREE, 1},
-  {"RANDOM_TREE_SCORE",             (DL_FUNC) &RANDOM_TREE_SCORE, 2},
   {"_TreeSearch_ts_test_strategy_tracker", (DL_FUNC) &_TreeSearch_ts_test_strategy_tracker, 2},
   {"_TreeSearch_ts_tbr_diagnostics", (DL_FUNC) &_TreeSearch_ts_tbr_diagnostics, 12},
   {"_TreeSearch_ts_ev_cache_key_probe", (DL_FUNC) &_TreeSearch_ts_ev_cache_key_probe, 9},